requirements.txt
pandas>=2.2.0
jinja2>=3.1.2
boto3>=1.34.0
botocore>=1.34.0
python-dotenv>=1.0.0
sentry-sdk>=1.40.0
google-analytics-data>=0.18.17
requests>=2.31.0
